PressureSensor: change unit of dummy pressure values
The dummy pressure values should be of double type just like the actual
values read from sensor.

+++ include/PressureSensor.h
@@ -63,7 +63,7 @@ private:
 	unsigned int totalNPWsDetected;
 	uint64_t npwBufferExpiryTime;   //NPW Buffer Expiry time in ms.
 
-	uint32_t readSensorValueDummy();
+	double readSensorValueDummy();
 
 	void fillCircularBufferWithDummyValues();
     void createNPWBuffer();
